+++ components/local-app/cmd/workspace-create.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/bufbuild/connect-go"
 	v1 "github.com/gitpod-io/gitpod/components/public-api/go/experimental/v1"
-	"github.com/gitpod-io/local-app/pkg/common"
 	"github.com/gitpod-io/local-app/pkg/config"
+	"github.com/gitpod-io/local-app/pkg/helper"
 	"github.com/spf13/cobra"
 )
 
@@ -73,16 +73,16 @@ var workspaceCreateCmd = &cobra.Command{
 			return nil
 		}
 
-		err = common.ObserveWorkspaceUntilStarted(ctx, gitpod, workspaceID)
+		err = helper.ObserveWorkspaceUntilStarted(ctx, gitpod, workspaceID)
 		if err != nil {
 			return err
 		}
 
 		if workspaceCreateOpts.StartOpts.OpenSSH {
-			return common.SshConnectToWs(ctx, gitpod, workspaceID, false)
+			return helper.SSHConnectToWorkspace(ctx, gitpod, workspaceID, false)
 		}
 		if workspaceCreateOpts.StartOpts.OpenEditor {
-			return common.OpenWorkspaceInPreferredEditor(ctx, gitpod, workspaceID)
+			return helper.OpenWorkspaceInPreferredEditor(ctx, gitpod, workspaceID)
 		}
 
 		return nil
